requirements.txt
# Core dependencies
click>=8.0.0
Pillow>=9.0.0
mutagen>=1.45.0
PyPDF2>=3.0.0
python-docx>=0.8.11

# Optional dependencies for enhanced functionality
pyyaml>=6.0
tabulate>=0.9.0
openpyxl>=3.0.0
chardet>=5.0.0

# Development dependencies (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=5.0.0
# mypy>=0.991